kubernetes/base/ai-analytics/requirements.txt
numpy==1.21.0
pandas==1.3.0
tensorflow==2.6.0
pyyaml==5.4.1
requests==2.26.0
python-dateutil==2.8.2
scikit-learn==0.24.2 